+++ src/main/java/brig/concord/ConcordIcons.java
@@ -7,4 +7,19 @@ import javax.swing.*;
 public final class ConcordIcons {
 
     public static final Icon FILE = IconLoader.getIcon("/icons/concord.png", ConcordIcons.class);
+
+    public static final Icon FLOWS = IconLoader.getIcon("/icons/structure-view/flows.svg", ConcordIcons.class);
+    public static final Icon FLOW = IconLoader.getIcon("/icons/structure-view/flow.svg", ConcordIcons.class);
+    public static final Icon FORMS = IconLoader.getIcon("/icons/structure-view/forms.svg", ConcordIcons.class);
+    public static final Icon FORM = IconLoader.getIcon("/icons/structure-view/form.svg", ConcordIcons.class);
+    public static final Icon RESOURCES = IconLoader.getIcon("/icons/structure-view/resources.svg", ConcordIcons.class);
+    public static final Icon IMPORTS = IconLoader.getIcon("/icons/structure-view/imports.svg", ConcordIcons.class);
+    public static final Icon IMPORT = IconLoader.getIcon("/icons/structure-view/import.svg", ConcordIcons.class);
+    public static final Icon CONFIGURATION = IconLoader.getIcon("/icons/structure-view/configuration.svg", ConcordIcons.class);
+    public static final Icon TRIGGERS = IconLoader.getIcon("/icons/structure-view/triggers.svg", ConcordIcons.class);
+    public static final Icon TRIGGER = IconLoader.getIcon("/icons/structure-view/trigger.svg", ConcordIcons.class);
+    public static final Icon PUBLIC_FLOWS = IconLoader.getIcon("/icons/structure-view/public-flows.svg", ConcordIcons.class);
+    public static final Icon PUBLIC_FLOW = IconLoader.getIcon("/icons/structure-view/public-flow.svg", ConcordIcons.class);
+    public static final Icon PROFILES = IconLoader.getIcon("/icons/structure-view/profiles.svg", ConcordIcons.class);
+    public static final Icon PROFILE = IconLoader.getIcon("/icons/structure-view/profile.svg", ConcordIcons.class);
 }